Refactor dismissible notice
