+++ js/src/components/dismissible-notice/index.js
@@ -3,6 +3,7 @@
  */
 import { Notice } from '@wordpress/components';
 import { useState } from '@wordpress/element';
+import { noop } from 'lodash';
 
 /**
  * Internal dependencies
@@ -16,14 +17,16 @@ import localStorage from '.~/utils/localStorage';
  *
  * @param {Object} props React props.
  * @param {JSX.Element} props.children Children to render.
- * @param {string} props.className Class to use in the Notice component.
  * @param {string|null} props.localStorageKey Local Storage Key where is keep the dismiss state.
+ * @param {Function} props.onRemove Callback when clicking on remove notice
+ * @param {Object} props.restProps Props to be forwarded to Notice Component. Like className.
  * @return {JSX.Element} {@link Notice} element with the info message and the link to the documentation.
  */
 const DismissibleNotice = ( {
 	children,
-	className,
 	localStorageKey = null,
+	onRemove = noop,
+	...rest
 } ) => {
 	const defaultDismissedValue = localStorageKey
 		? !! localStorage.get( localStorageKey )
@@ -31,10 +34,12 @@ const DismissibleNotice = ( {
 
 	const [ isDismissed, setIsDismissed ] = useState( defaultDismissedValue );
 
-	const onRemove = () => {
+	const handleRemove = () => {
 		if ( localStorageKey ) {
 			localStorage.set( localStorageKey, true );
 		}
+
+		onRemove();
 
 		setIsDismissed( true );
 	};
@@ -44,7 +49,7 @@ const DismissibleNotice = ( {
 	}
 
 	return (
-		<Notice className={ className } onRemove={ onRemove }>
+		<Notice onRemove={ handleRemove } { ...rest }>
 			{ children }
 		</Notice>
 	);
